components/python/client/requirements.txt
# primary dependencies

paho-mqtt

# secondary dependencies
